restructure unittests files. Add add_document for optional
