+++ api/coverage.txt
@@ -1,24 +1,22 @@
 Name                                                       Stmts   Miss    Cover   Missing
 ------------------------------------------------------------------------------------------
-utils/form_to_schema.py                                       29     22    24.1%   10-46, 50
-core/use_case/utils/get_ui_recipe.py                          23     17    26.1%   9-19, 24-29
+utils/form_to_schema.py                                       28     22    21.4%   9-45, 49
 core/use_case/add_file_use_case.py                            45     33    26.7%   21-28, 32-52, 66, 69-86
 classes/data_source.py                                        22     16    27.3%   9-12, 15-27
 core/utility.py                                               71     50    29.6%   22-36, 40-48, 53-58, 62-66, 71-76, 84-90, 93-97
 tests/core/use_case/test_add_file_to_package_use_case.py      20     14    30.0%   11-33
 core/repository/__init__.py                                   34     22    35.3%   11-13, 16-21, 24-33, 36-42, 45, 48
-classes/ui_recipe.py                                          33     20    39.4%   13-14, 19-21, 27-48, 53
-classes/recipe.py                                             40     24    40.0%   15-16, 21-25, 28-30, 33-34, 37-50, 55
 utils/helper_functions.py                                     15      9    40.0%   6-9, 13-17
 tests/core/domain/test_package.py                             24     14    41.7%   8-12, 17-19, 24-26, 31-34
 utils/data_structure/compare.py                               28     16    42.9%   9-10, 18-23, 25-32
 core/repository/mongo/__init__.py                             30     17    43.3%   13-22, 25-26, 29-33, 36-40, 43, 46, 49
+classes/recipe.py                                             39     19    51.3%   20, 35-55, 58, 66
 tests/classes/test_tree_error_node.py                         15      7    53.3%   29-45
 utils/data_structure/find.py                                  37     16    56.8%   23-24, 28-30, 34-45
 core/repository/repository_exceptions.py                      15      6    60.0%   3-4, 7, 12, 17, 22
 core/shared/use_case.py                                       13      5    61.5%   8, 11-13, 16
 core/repository/repository_factory.py                          8      3    62.5%   8-10
-classes/blueprint.py                                          55     20    63.6%   32-36, 49, 60, 63-67, 70, 76, 79-85
+classes/blueprint.py                                          58     20    65.5%   36-40, 53, 64, 67-71, 74, 80, 83-89
 classes/dto.py                                                48     16    66.7%   19, 23-26, 29-33, 40, 44-45, 53, 57-58, 65
 core/use_case/create_application_use_case.py                 108     36    66.7%   172, 175, 185, 193-194, 200-228, 232-236, 248, 270-272
 core/shared/response_object.py                                37     12    67.6%   20-21, 24-26, 30, 33, 37, 41, 45, 49-52
@@ -26,9 +24,9 @@ core/repository/file/__init__.py                              32      8    75.0%
 utils/data_structure/traverse.py                              35      8    77.1%   20, 53-55, 59-62
 tests/core/repository/mongo/test_document_repository.py        9      2    77.8%   12-13
 utils/get_data_type.py                                         9      2    77.8%   11-12
-classes/tree_node.py                                         308     65    78.9%   102-108, 132-134, 175-176, 182, 195-196, 203-206, 209, 252, 255-257, 263, 266, 275-278, 293, 299, 311-344, 363, 371, 385, 395, 446-447, 461-462, 472-473, 489-495, 516, 524
+classes/tree_node.py                                         310     64    79.4%   102-108, 132-134, 175-176, 182, 194-195, 202-205, 251, 254-256, 262, 265, 274-277, 292, 298, 311-344, 363, 371, 385, 395, 447-448, 462-463, 473-474, 490-496, 518, 526
 core/shared/request_object.py                                 17      3    82.4%   6, 12, 20
-core/service/document_service.py                             144     23    84.0%   17, 28, 50, 54-57, 71-72, 104, 107, 127-132, 139, 148, 153, 168, 193, 205, 211, 226, 232
+core/service/document_service.py                             149     23    84.6%   17, 28, 50, 54-57, 71-72, 104, 107, 127-132, 139, 148, 153, 168, 193, 205, 211, 225, 231
 classes/dimension.py                                          38      5    86.8%   16, 20, 38, 53, 58
 services/database.py                                           8      1    87.5%   10
 core/use_case/utils/create_entity.py                          76      8    89.5%   12-13, 16, 21, 46-48, 56
@@ -37,9 +35,11 @@ core/enums.py                                                 43      4    90.7%
 classes/blueprint_attribute.py                                33      3    90.9%   36, 47, 61
 tests/utils/test_create_default_array.py                      47      4    91.5%   95, 97, 99, 157
 utils/group_by.py                                             17      1    94.1%   18
-tests/core/service/test_document_service.py                  271      8    97.0%   112, 127, 159, 198, 229, 260, 299, 597
+tests/core/document_service/test_remove.py                    79      4    94.9%   23, 55, 94, 125
+tests/core/document_service/test_get.py                       32      1    96.9%   44
+tests/core/document_service/test_rename.py                    83      2    97.6%   31, 70
 tests/classes/test_tree_node.py                              182      1    99.5%   119
 ------------------------------------------------------------------------------------------
-TOTAL                                                       2354    544    76.9%
+TOTAL                                                       2390    500    79.1%
 
-25 files skipped due to complete coverage.
+28 files skipped due to complete coverage.
