+++ api/tests/core/document_service/test_update.py
@@ -1,0 +1,151 @@
+import unittest
+from unittest import mock
+
+from classes.dto import DTO
+from classes.tree_node import Node
+from core.repository import Repository
+from core.service.document_service import DocumentService
+from tests.core.document_service.common import blueprint_provider
+from utils.data_structure.compare import pretty_eq
+
+
+class DocumentServiceTestCase(unittest.TestCase):
+    def test_update_single_optional_complex(self):
+        repository: Repository = mock.Mock()
+
+        doc_storage = {
+            "1": {
+                "_id": "1",
+                "name": "Parent",
+                "description": "",
+                "type": "blueprint_with_optional_attr",
+                "im_optional": {},
+            }
+        }
+
+        doc_1_after = {
+            "_id": "1",
+            "name": "Parent",
+            "description": "Test",
+            "type": "blueprint_with_optional_attr",
+            "im_optional": {},
+        }
+
+        def mock_get(document_id: str):
+            return DTO(doc_storage[document_id])
+
+        def mock_update(dto: DTO):
+            doc_storage[dto.uid] = dto.data
+            return None
+
+        def repository_provider(data_source_id):
+            if data_source_id == "testing":
+                return repository
+
+        repository.get = mock_get
+        repository.update = mock_update
+        document_service = DocumentService(
+            blueprint_provider=blueprint_provider, repository_provider=repository_provider
+        )
+
+        node: Node = document_service.get_by_uid("testing", "1")
+        node.update(doc_1_after)
+        document_service.save(node, "testing")
+
+        assert pretty_eq(doc_1_after, doc_storage["1"]) is None
+
+    def test_add_optional(self):
+        repository: Repository = mock.Mock()
+
+        doc_storage = {
+            "1": {
+                "_id": "1",
+                "name": "Parent",
+                "description": "",
+                "type": "blueprint_with_optional_attr",
+                "im_optional": {},
+            }
+        }
+
+        doc_1_after = {
+            "_id": "1",
+            "name": "Parent",
+            "description": "",
+            "type": "blueprint_with_optional_attr",
+            "im_optional": {"name": "new_entity", "type": "blueprint_2", "description": "This is my new entity"},
+        }
+
+        def mock_get(document_id: str):
+            return DTO(doc_storage[document_id])
+
+        def mock_update(dto: DTO):
+            doc_storage[dto.uid] = dto.data
+            return None
+
+        def repository_provider(data_source_id):
+            if data_source_id == "testing":
+                return repository
+
+        repository.get = mock_get
+        repository.update = mock_update
+        document_service = DocumentService(
+            blueprint_provider=blueprint_provider, repository_provider=repository_provider
+        )
+        document_service.add_document(
+            "testing", "1", "blueprint_2", "new_entity", "This is my new entity", "im_optional"
+        )
+
+        assert pretty_eq(doc_1_after, doc_storage["1"]) is None
+
+    def test_add_optional_nested(self):
+        repository: Repository = mock.Mock()
+
+        doc_storage = {
+            "1": {
+                "_id": "1",
+                "name": "Parent",
+                "description": "",
+                "type": "blueprint_with_nested_optional_attr",
+                "nested_with_optional": {
+                    "name": "Parent",
+                    "description": "",
+                    "type": "blueprint_with_optional_attr",
+                    "im_optional": {},
+                },
+            }
+        }
+
+        doc_1_after = {
+            "_id": "1",
+            "name": "Parent",
+            "description": "",
+            "type": "blueprint_with_nested_optional_attr",
+            "nested_with_optional": {
+                "name": "Parent",
+                "description": "",
+                "type": "blueprint_with_optional_attr",
+                "im_optional": {"name": "new_entity", "type": "blueprint_2", "description": "This is my new entity"},
+            },
+        }
+
+        def mock_get(document_id: str):
+            return DTO(doc_storage[document_id])
+
+        def mock_update(dto: DTO):
+            doc_storage[dto.uid] = dto.data
+            return None
+
+        def repository_provider(data_source_id):
+            if data_source_id == "testing":
+                return repository
+
+        repository.get = mock_get
+        repository.update = mock_update
+        document_service = DocumentService(
+            blueprint_provider=blueprint_provider, repository_provider=repository_provider
+        )
+        document_service.add_document(
+            "testing", "1", "blueprint_2", "new_entity", "This is my new entity", "nested_with_optional.im_optional"
+        )
+
+        assert pretty_eq(doc_1_after, doc_storage["1"]) is None
